chore(ui): rename the `day_dividers` module to `date_dividers`
