+++ crates/matrix-sdk-ui/src/timeline/controller/state.rs
@@ -53,7 +53,7 @@ use super::{
 use crate::{
     events::SyncTimelineEventWithoutContent,
     timeline::{
-        day_dividers::DateDividerAdjuster,
+        date_dividers::DateDividerAdjuster,
         event_handler::{
             Flow, HandleEventResult, TimelineEventContext, TimelineEventHandler, TimelineEventKind,
             TimelineItemPosition,
