+++ crates/matrix-sdk-ui/src/timeline/mod.rs
@@ -59,7 +59,7 @@ use crate::timeline::pinned_events_loader::PinnedEventsRoom;
 
 mod builder;
 mod controller;
-mod day_dividers;
+mod date_dividers;
 mod error;
 mod event_handler;
 mod event_item;
